Relatively massive update mostly related to increasing test coverage.  Most modules are in the 75-100% range with most of the issues being exception cases and my gratuitous creation of helper functions.
Also,this marks the conversion of the project from VSCode to PyCharm.  This was an experiment for me to use VSCode on a large project and I just found it to be flakey around testing.  It just stopped recoginziing all tests.  At work this stuff never happens so I bought the individual version of pycharm and everyting worked out of the box with minimal fuss.